[Refactor] Rename Collect2 > Collect
