+++ src/Apps/Collect/Routes/Collection/Components/Header/FeaturedArtists.tsx
