+++ src/Apps/Collect/Routes/Collection/Components/Header/__tests__/fixtures/artworks.ts
