+++ src/Apps/Collect/Routes/Collection/index.tsx
@@ -1,8 +1,8 @@
 import { Box, Separator } from "@artsy/palette"
 import { Collection_collection } from "__generated__/Collection_collection.graphql"
-import { SeoProductsForArtworks } from "Apps/Collect2/Components/SeoProductsForArtworks"
-import { SeoProductsForCollections } from "Apps/Collect2/Components/SeoProductsForCollections"
-import { CollectionFilterFragmentContainer as CollectionHeader } from "Apps/Collect2/Routes/Collection/Components/Header"
+import { SeoProductsForArtworks } from "Apps/Collect/Components/SeoProductsForArtworks"
+import { SeoProductsForCollections } from "Apps/Collect/Components/SeoProductsForCollections"
+import { CollectionFilterFragmentContainer as CollectionHeader } from "Apps/Collect/Routes/Collection/Components/Header"
 import { AppContainer } from "Apps/Components/AppContainer"
 import { track } from "Artsy/Analytics"
 import * as Schema from "Artsy/Analytics/Schema"
