+++ src/Apps/Collect/Routes/Collections/__tests__/CollectionsApp.test.tsx
@@ -1,8 +1,8 @@
 import { EntityHeader } from "@artsy/palette"
 import { CollectionsAppTestQueryRawResponse } from "__generated__/CollectionsAppTestQuery.graphql"
 import { CategoriesFixture } from "Apps/__tests__/Fixtures/Collections"
-import { CollectionsAppFragmentContainer as CollectionsApp } from "Apps/Collect2/Routes/Collections"
-import { CollectionsGrid } from "Apps/Collect2/Routes/Collections/Components/CollectionsGrid"
+import { CollectionsAppFragmentContainer as CollectionsApp } from "Apps/Collect/Routes/Collections"
+import { CollectionsGrid } from "Apps/Collect/Routes/Collections/Components/CollectionsGrid"
 import { BreadCrumbList } from "Components/Seo"
 import { MockBoot, renderRelayTree } from "DevTools"
 import React from "react"
