+++ src/Apps/Collect/Utils/urlBuilder.ts
